Unbreak debian/rules autogen.sh call for cross-compilation.
autogen.sh calls './configure "$@"' as a last command, while debian/rules does
not pass any arguments to it, hence configure fails under dpkg-buildpackage
-a<arch> -- CC points to the cross-compiler, but --host and --build are not
provided to configure, so it tries to use cross-compiler as a native compiler.

To avoid this, just pass all configure arguments to the autogen.sh and remove
separate configure call.